[Headless] Fix forced colors test suite for headless shell
This CL addresses the issue with the wpt/forced-colors-mode test suite,
which was failing when executed in the headless shell. The failure
occurred because the forced colors mode was not being enabled in
headless shell via the flag.

To resolve this, the logic to update the forced colors pref has been
added to HeadlessContentBrowserClient::OverrideWebkitPrefs(),similar to
the implementation in ShellContentBrowserClient::OverrideWebkitPrefs().

Bug: 368396826
Change-Id: I56d3e9b77750e4c71ee963586cb179ff676c642f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5939264
Reviewed-by: Peter Kvitek <kvitekp@chromium.org>
Commit-Queue: Sam Davis Omekara <samomekarajr@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1371178}