[NSP] Set request initiator for reporting preflights and add comments
This updates the Reporting service so that the preflights it sends have
the request initiator field set. I don't think this introduces any
changes in practice since credentials are disabled for these requests,
but it seems worthwhile for consistency in indicating that these are not
browser-initiated navigations (and aligns these with preflight requests
constructed elsewhere [1]).

Also, StartPayloadRequest used to set the initiator using the
IsolationInfo frame_origin, but this was changed to using the
upload->request_origin because for Network State Partitioning we were
experimenting with removing the frame_origin field from IsolationInfo
[2]. I was wondering whether we should switch back now that the
IsolationInfo frame_origin is here to stay. It seems that using
upload->request_origin is actually what we want, though, so I've tried
to add comments that clarify this more. I've also added a CHECK for
the only fact I could identify about the IsolationInfo used.

[1] https://source.chromium.org/chromium/chromium/src/+/main:services/network/cors/preflight_controller.cc;l=199;drc=5dcd503d1328cb5410733acd073e547e2df0b079

[2] https://chromium-review.googlesource.com/c/chromium/src/+/3442039/20/net/reporting/reporting_uploader.cc#199

Bug: None
Change-Id: Ifd7f21520ae8c2c34c82c7c4958fd338f9eace16
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5899836
Commit-Queue: Andrew Williams <awillia@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1371183}