[//content] Rename variable in browser_main_loop.cc
Opportunistic cleanup: `established_gpu_channel` is a misleading name
for this variable as it gets passed to a function wherein `true` means
that the function *should establish* the GPU channel. Rename the
variable and clarify the one place where it currently is used in a
way that checks the variable to determine whether the GPU process was
already launched.

Change-Id: I946c9b950e6a76afa10530fb2919c5cc468ff3aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5938090
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1371212}