[Gardening] /gpu/fast/canvas/webgl/create-query-crash.html is flaky
The test recently started failing on graphite-enabled test suites
on mac-skia-alt-arm64. The cause for the failure is an additional,
unexpected line in the test output:

+CONSOLE WARNING: [.WebGL-0x12c001d6200]GL Driver Message (OpenGL,
Performance, GL_CLOSE_PATH_NV, High): GPU stall due to ReadPixels

This already seems to be happening to a few other tests - this CL
adds this WPT to the existing group with such "failures".

Change-Id: Ie7bed4c5015ead4e0a7f6b2dad89a15b178af34e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5946244
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: Jan Keitel <jkeitel@google.com>
Cr-Commit-Position: refs/heads/main@{#1371276}