[Android][KA] Add suggestion type to the UserInfoField.
This CL moves AccessorySuggestionType enum from metric to the
accessory_data.h. This enum is now added to the AccessorySheetField
on the C++ side and to the UserInfoField on the Java side.

AccessoryTabType was used mainly as a substitute of this enum before.
With the addition of plus addresses suggestions, it's not possible
to determine the suggestion type by the tab type. So this CL
encapsulates this information in the UserInfoField data class.

Next steps: refactor the metric logging to accept the
AccessorySuggestionType directly.

Bug: 327838324
Change-Id: Ic5a81f0a8fa52180bcfdc8ed652db5e80d784c87
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5939936
Commit-Queue: Timofey Chudakov <tchudakov@google.com>
Reviewed-by: Friedrich Hauser <friedrichh@chromium.org>
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1371292}