Forcing pseudo classes forces off dual classes
Currently we provide a way to force pseudo classes to be on, but not off. This is particularly weird with dual pairs like :enabled/:disabled - the functionality can only be used to get an element to render as simultaneously both. With this change, forcing one of a dual pair will force the complement off.

Bug: 371926664
Change-Id: I97ef0cca2c83cfd2b344383630bbcd07b4de6c8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5938236
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Auto-Submit: Eric Leese <leese@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1371299}