Fix flaky test on win-asan
The test was flaky because the initial browser was sometimes missing. We don't really need that browser, so if the guest browser works, then the test might still succeed.

This CL removes the check for the initial browser and the test should be
no longer flaky.

Bug: 373890422
Change-Id: I7eae7ddab449dd55115d5d0148b0bb0895f0ecb6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5945785
Commit-Queue: Nicolas Dossou-Gbété <dgn@chromium.org>
Auto-Submit: Christian Dullweber <dullweber@chromium.org>
Reviewed-by: Nicolas Dossou-Gbété <dgn@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1371406}