[SAH] Deflake request header related browsertests
The previous approach saved only the most recent set of request
headers, when issuing fetches and loading pages. This assumes that all
of the browser's fetches occur in a predictable order; but they don't.
This meant that these tests are flaky, since the test's fetches are
racing with each other.

This CL removes that assumption by saving *all* of the observed sets
of request headers, and being more lax about the expected ordering. In
exchange, the tests can be more descriptive about the full set of
headers that should be observed over the lifetime of the test.

Bug: 332335089
Change-Id: Ib4f97e6cc33bacd0286cf611a918961d596eeb47
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5942921
Commit-Queue: Chris Fredrickson <cfredric@chromium.org>
Reviewed-by: Ari Chivukula <arichiv@chromium.org>
Auto-Submit: Chris Fredrickson <cfredric@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1371439}