[Payments] Log sqlite errors instead of crashing
Before this patch, failures in web payments sqlite table creation and
alteration had a NOTREACHED() macro, because only SQL syntax (which
developers can control) was expected to be a possible failure reason.
However, the list of possible failure reasons in
//sql/sqlite_result_code_values.h shows that sqlite can fail for a large
number of reasons that are not under developer control, such as lack of
memory, full disk, disk corruption, and so on. This was not noticed
until NOTREACHED() started to become CHECK():
https://groups.google.com/a/chromium.org/g/chromium-dev/c/7PuBcFK2l1s/m/lFCar7pfAQAJ

This patch changes the NOTREACHED() macros to LOG(ERROR) macros in the
payments sqlite table creation and alteration.

After this patch, failures in payments table creation and alteration are
logged (locally) and handled by the calling code.

Note:
There is an ongoing discussion of the best way to improve handling of
sqlite errors, e.g., should SPC enrollment fail earlier, when it is
known ahead of time that SPC won't be able to save the credential
identifier in the table?

Fixed: 40947701
Change-Id: I6dc44a53733e1fb356fa8695e033257485c8ff78
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5939089
Commit-Queue: Rouslan Solomakhin <rouslan@chromium.org>
Reviewed-by: Slobodan Pejic <slobodan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1371462}