Improve word match recall by trimming tokens & treating like query terms
This CL adds string view trimming logic to the word match boost search
so that terms next to various ASCII characters won't be missed. The test
data is updated to show an example where a word in quotation marks was
missed before but now is found to increase the boost.  Queries and
passages are also now tokenized and trimmed in exactly the same way,
with shared logic and constants.

Bug: 366208372, 361654005
Change-Id: I634a1a19e7a5afedf4378994c5e5eac159c75589
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5941613
Reviewed-by: Orin Jaworski <orinj@chromium.org>
Reviewed-by: Roman Arora <romanarora@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Orin Jaworski <orinj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1371546}