Default the package destination path to the proto file name.
The "dest_path" was used to set the CIPD package install location, which
should be unique for each avd config so that packages in each AVD will
be installed in separate dirs.

Given the proto file name is guaranteed to be unique under the same dir,
it is a good fit to be the fallback value for the "dest_path"

This will make it easier and more straightforward to manage the packages
in the avd config.

Should be a no-op CL as the existing "dest_path" values are already
the same as the proto file names

Bug: 364943269
Change-Id: I317c9fcbddd972e7d6157b9b2643bc7727ba130b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5943569
Commit-Queue: Haiyang Pan <hypan@google.com>
Reviewed-by: Renuga Nanjappan <rnanjappan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1371563}