[views-ax] Migrating text offset related attributes in Label
This CL migrates the text offset related accessibility attributes in
Label to use the new "push" system for Views where we cache
the accessible values. As part of this system, we introduced
functionality that will enable Views authors to defer the computation
of expensive accessibility attributes to only be computed when
accessibility is on for Views. These attributes will not be computed
or cached until accessibility is on, once it is we will first
compute the values at the moment when accessibility turns on, and
then will be recomputed whenever the value should change.

This CL is part of the ViewsAX project:
https://docs.google.com/document/d/1Ku7HOyDsiZem1yaV6ccZ-tz3lO2XR2NEcm8HjR6d-VY/edit#heading=h.ke1u3utej413

Bug: 325137417
Change-Id: I8b9ecdddd226aef8e61e0cbb057a8d019aa2d675
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5888850
Reviewed-by: Dirk Pranke <dpranke@google.com>
Commit-Queue: Javier Contreras <javiercon@microsoft.com>
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1371573}