[freezing] Discard frozen browsing instances with growing private memory.
A frozen page cannot execute tasks and therefore typically doesn't
grow memory usage. However, there exist cases for which its memory
usage can grow, for example if another page or worker sends messages
to it. In that case, we'll discard the frozen page with growing
memory (ignoring all discard opt-outs, because the tab is already
frozen and thus not providing useful functionality, and user state
loss is considered better than unbounded memory usage growth). This
intervention will be coupled with developer messaging to discourage
sending a lot of messages to a frozen tab.

This intervention is implemented in 2 CLs:

1. (crrev.com/c/5927211) Add a new discard reason:
   FROZEN_WITH_GROWING_MEMORY.
2. (this CL) Discard a page with FROZEN_WITH_GROWING_MEMORY when the
   private memory footprint of a tab grows above a threshold.

Bug: 325954772
Change-Id: I7a9dc82bfab485ef5e25fc1026d38225b5173bbb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5926052
Commit-Queue: Francois Pierre Doray <fdoray@chromium.org>
Reviewed-by: Joe Mason <joenotcharles@google.com>
Cr-Commit-Position: refs/heads/main@{#1371624}