Replace SnackbarView#announceForAccessibility with #setAccessibilityPaneTitle
announceForAccessibility is discouraged and should be used sparingly.
SnackbarView currently uses this API to announce the appearance of the
Snackbar. This can be replaced with #setAccessibilityPaneTitle, which
matches what the Material widget does.

SnackbarView also provides APIs for announcing confirmation of the
Snackbar action. This will be addressed in a separate CL.

Bug: 374789095
Change-Id: Iea198044f13711c8914cb129ecc5a25630b49eab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5949403
Reviewed-by: Lijin Shen <lazzzis@google.com>
Commit-Queue: Theresa Sullivan <twellington@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1371634}