Clean up Tab Observers for warmed up/prerendered tabs.
The HiddenTabHolder observer hasn't owned the WindowAndroid since the
tab creation logic was moved into WarmupManager, so it no longer needs
to destroy the WindowAndroid and that can be cleaned up.

Also, when the warmed up or prerendered tab crashed, it could
potentially cause a different warmed up tab to be destroyed. In
practice the prerendered tab is re-attached immedaitely after being
taken so destroying the 'current' prerendered tab wasn't actually an
issue.

However, we could get into a situation where a prerendered tab crashing
causes an unrelated warmed up tab to be destroyed. This is probably
very rare, as only ~1% of warmed up tabs are reported to crash (or have
their renderer killed) before being used.

Bug: 374263402
Change-Id: Ibe39f3ffdeac3ff464d16236a3174c277f846dbf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5940710
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Commit-Queue: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1371644}