[Suggestion][Clank] Add network in main text content description
The logic for announcing the network of nicknamed cards is currently
handled in Java. A `NETWORK_NAME` property receives the value from
the credit card object, which is then combined with the card name for
announcement.
However, a new approach is being implemented with the suggestion.h
object. Instead of handling the network name separately in Java, a
`main_text_content_description` attribute is defined on the native
side. This attribute includes the network information if it differs
from the card name. This value is then passed to Java and assigned
to the `main_text` content description.

Video (Identical before and after):
Before:
http://screencast/cast/NDk1MDAzMjU5ODk1ODA4MHw1MDRmN2QxMC00Zg
After:
http://screencast/cast/NTg3MjAzMzYzMTA0MzU4NHw1ZTYzNWZjNS02Yg

Bug: 360440916
Change-Id: Ibc4532ccdd82743c47b3e9c965bd11ce0eadc2bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5921236
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Reviewed-by: Olivia Saul <jsaul@google.com>
Commit-Queue: Viplav Kadam <viplavkadam@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1371665}