[WebAuthn] Check for conditional UI being called before DOMContentLoaded
This adjusts a metric that records the time between DOMContentLoaded
and a WebAuthn conditional UI get request being invoked.

Some scripts execute before the DOMContentLoaded event occurs, and in
that case we are computing the time from a null TimeTicks value, which
is wrong. This changes the metric to always emit zero in the situation.

Change-Id: Ia6a63ebaef5e91ceea8d1259b9a67e77b332eae3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5950004
Commit-Queue: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Nina Satragno <nsatragno@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1371668}