[CrOS Networking] Avoid responding if javascript calls are not allowed
When user attempts to generate logs from chrome://network and refreshes
the page when logs are still being generated, a crash occurs. This
happens due to a failed Check() where javascript calls are not allowed
(as the page isn't available).

This change adds the check before responding to handle this case.

Bug: b/374223603
Change-Id: Ic7530151380c0575dc2e48b494607364e2541105
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5946927
Commit-Queue: Nikhil Nayunigari <nikhilcn@google.com>
Reviewed-by: Chad Duffin <chadduffin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1371716}