[Compare] Guard entry point controller creation behind flag
This CL adds a feature flag check before creating the entry point
controller. By doing so, ProductSpecificationsButton can no longer
assume that the controller is always available. Even though the
creation was not behind flag, all the logics of the controller are
essentially behind flag before this CL, so we expect no user-visible
change.

This CL also reverts https://crrev.com/c/5836137 which is a
temporary fix to be merged back, as the creation is conditioned on
regular profile now.

Bug: 363166165
Change-Id: I12898145342f96f5cac14411f8e45e5f3029fdf0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5939396
Reviewed-by: Erik Chen <erikchen@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Commit-Queue: Yue Zhang <yuezhanggg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1371719}