trusted_vault: remove obsolete code
This removes several things from //components/trusted_vault that were
originally added to support GPM passkeys on ChromeOS. ChromeOS has since
launched support for GPM passkeys using the enclave authenticator,
making some of this trusted_vault code obsolete.

First is the instantiation of a TrustedVaultClient for the passkeys
security domain. The enclave authenticator never stores domain secrets
locally on disk.

Second are the RecoveryStoreController and RecoveryKeyProvider which
together were intended to upload ChromeOS LSKFs to the recovery key
store service. The enclave authenticator doesn't use the LSKF for
recovery.  Note that RecoveryKeyStoreConnection is still actively used
by the GPM enclave authenticator for uploading the GPM PIN, so it
remains.

All of the deleted code was gated behind the `device::kChromeosPasskeys`
feature flag which was never enabled by default and is about to be
deleted as well.

(Also fix a missing dependency in `feedback_private` on chromeos that
otherwise causes the build to fail after the code removal in this CL.)

Change-Id: I2e78b4e675a5958051c68e5e6d5de40042349832
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5941506
Reviewed-by: Xiangdong Kong <xiangdongkong@google.com>
Commit-Queue: Xiangdong Kong <xiangdongkong@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Auto-Submit: Martin Kreichgauer <martinkr@google.com>
Cr-Commit-Position: refs/heads/main@{#1371721}