[HFM] Disable heuristics if device or profile is managed
This CL disables Site Engagement and Typically Secure User heuristics
if the device or profile is enterprise managed. It also adds checks to
ensure that we never enable HFM on these devices even if a previous
version of Chrome added a domain to the HTTPS enforcement list.

The CL uses an existing function to check whether the device is
enterprise managed. The function was incomplete on ChromeOS, so it
also fixes that.

Bug: 374334530
Change-Id: I664aa2eb63c410291adc4ace747d9f0fa744b57b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5938490
Reviewed-by: Chris Thompson <cthomp@chromium.org>
Commit-Queue: Mustafa Emre Acer <meacer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1371763}