Fix null dereference in TlsStreamAttempt
In TlsStreamAttempt::DoTlsAttemptComplete(), `ssl_socket_` is not
always valid because DoTlsAttempt() may return ERR_ABORTED without
creating a SSLClientSocket. Change
SSLClientSocket::RecordSSLConnectResult() to take a pointer instead
of a reference and only use it when it's not null.

Confirmed that
HttpStreamPoolAttemptManagerTest.EchSvcbReliantQuicOnlyAbortTcpAttempt
passed with this CL for a ubsan build.

Bug: 374797733
Change-Id: Ia7a43a7f7d94ffc770d0fd9c7e56f5ab0095e17c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5951123
Commit-Queue: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: David Benjamin <davidben@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1371776}