Implement Prefetch matching
Add the PrefetchMatches() function which checks that a ResourceRequest
from a render process is a sufficiently good match for a ResourceRequest
that was previously used for a prefetch to be able to use the prefetch.
Also has debug-level diagnostics of exactly what went wrong in the event
of a mismatch. Add a StreamEnumValueTo() function to make the
diagnostics more readable.

As there are many fields and new ones are added and removed all the
time, a significant amount of metaprogramming is used to make this as
simple as possible.

The new code is not used in this CL in order to keep the size down.

BYPASS_LARGE_CHANGE_WARNING=Have to compare a lot of fields
NO_IFTTT=IFTTT seems to be mis-triggering. enums.xml was amended.

BUG=342445996

Change-Id: Icc822ad8fa5ab66a7277f96884001fa29e035391
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5931723
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Nidhi Jaju <nidhijaju@chromium.org>
Commit-Queue: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1371793}