[Extensions] Expand message tracker to track non-SW contexts.
Before this change the tracker didn't have a catch-all for messages
that go to destinations that are not a service worker. This is
necessary because there are places during message tracking where it is
difficult to determine what specific kind of destination type is being
targeted.

This is fine though because these are only for comparison to service
workers so them being less granular is okay.

This also makes some adjustments to naming as a result.

Bug: 371011217
Change-Id: I8639d5a1ab351b941525ca0a284ac488340e5b0b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5941000
Reviewed-by: Emilia Paz <emiliapaz@chromium.org>
Auto-Submit: Justin Lulejian <jlulejian@chromium.org>
Commit-Queue: Justin Lulejian <jlulejian@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1371809}