[a11y] Improves a11y spelling annotations behavior for UIA
Upon selecting words in text with spelling/grammar errors,
accessibility annotations for the same do not work once selection
overflows the misspelled literal by even a single character.

This CL fixes the same.
In the text selection, the code was setting the attribute value
for Spelling/Grammar as UNKNOWN if in the selection say there is
Spelling error for partial selection and then no annotation for
the rest of the selected text. After this CL, the code will dish
out SpellingError annotation if the partial text selection has that
error but the rest of the selection has no other annotation. Same
for GrammarError.

The bug linked below explains the issue in detail with a video demo

Bug: 364795299
Change-Id: I19cf8e64ac476e7a0d6a1078af288341ae062690
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5839087
Commit-Queue: Vinay Singh <vinaysingh@microsoft.com>
Reviewed-by: Divyansh Mangal <dmangal@microsoft.com>
Auto-Submit: Vinay Singh <vinaysingh@microsoft.com>
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1371919}