[Payments] Handle unknown HTTP status code for manifest downloads
Before this patch, downloading a payment manifest could return an
unknown HTTP status code, which would hit a NOTREACHED() when attempting
to convert that code into a human-readable error message for the web
developer.

This patch handles the case of an unknown status code. The string
"Unknown" is used as the human-readable reason for the download failure.

After this patch, an unknown HTTP status code for payment manifest
download does not hit a NOTREACHED().

Fixed: 329106312
Change-Id: Ie75fefbfb60b65aa728121c919625ab1c9acdbde
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5937915
Reviewed-by: Kavita Soni <kavitasoni@chromium.org>
Commit-Queue: Rouslan Solomakhin <rouslan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1371956}